+++ src/utils/channel_name_validation.rs
@@ -1,4 +1,3 @@
-use axum::http::StatusCode;
 use regex::Regex;
 
 const VALID_CHANNEL_NAME_REGEX: &str = r#"^(?:/[a-zA-Z0-9\-_!~()$@]+)+(?:(?:/\*\*)|(?:/\*)|/)?$"#;
@@ -16,60 +15,12 @@ impl Default for ChannelNameValidator {
 
 impl ChannelNameValidator {
     #[inline(always)]
-    pub(crate) fn validate(&self, name: &str) -> Result<(), StatusCode> {
-        if self.0.is_match(name) {
-            Ok(())
-        } else {
-            Err(StatusCode::BAD_REQUEST)
-        }
+    pub(crate) fn validate(&self, name: &str) -> bool {
+        self.0.is_match(name)
     }
-}
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    #[rustfmt::skip]
-    fn test_channel_name_validator() {
-        let validator = ChannelNameValidator::default();
-
-        for (channel, result) in [
-            (r#"/first1"#, Ok(())),
-            (r#"/first1*"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1**"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/*"#, Ok(())),
-            (r#"/first1/**"#, Ok(())),
-            (r#"/first1/"#, Ok(())),
-            (r#"/first1/*"#, Ok(())),
-            (r#"/first1/**"#, Ok(())),
-            (r#"/first1/second2"#, Ok(())),
-            (r#"/first1/second2*"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2**"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/"#, Ok(())),
-            (r#"/first1/second2/*"#, Ok(())),
-            (r#"/first1/second2/**"#, Ok(())),
-            (r#"/first1/second2/third3"#, Ok(())),
-            (r#"/first1/second2/third3*"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/third3**"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/third3/"#, Ok(())),
-            (r#"/first1/second2/third3/*"#, Ok(())),
-            (r#"/first1/second2/third3/**"#, Ok(())),
-            (r#"/first1/*/third3"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/*/third3/"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/*/third3/*"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/*/third3/**"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/**/"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/**/*"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/**/**"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/third3/-_!~()$@"#, Ok(())),
-            (r#"/first1/second2/third3/-_!~()$@*"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/third3/-_!~()$@**"#, Err(StatusCode::BAD_REQUEST)),
-            (r#"/first1/second2/third3/-_!~()$@/"#, Ok(())),
-            (r#"/first1/second2/third3/-_!~()$@/*"#, Ok(())),
-            (r#"/first1/second2/third3/-_!~()$@/**"#, Ok(())),
-        ] {
-            assert_eq!(validator.validate(channel), result, "{channel}");
-        }
+    #[inline(always)]
+    pub(crate) fn validate_error<E>(&self, name: &str, error: E) -> Result<(), E> {
+        self.validate(name).then_some(()).ok_or(error)
     }
 }