+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Add
 
 - Channel name validation.
+- Validate channel on `send` and `send_to_client`.
 
 ### Change